+++ types/module/mock_appmodule_test.go
@@ -28,3 +28,8 @@ type CoreAppModule interface {
 	appmodule.HasPrecommit
 	appmodule.HasPrepareCheckState
 }
+
+type CoreUpgradeAppModule interface {
+	CoreAppModule
+	appmodule.UpgradeModule
+}